api/requirements.txt
Flask==1.1.2
Markdown==3.2.1
Flask-RESTful==0.3.8
mysql-connector-python==8.0.19
rsa==4.0
pycryptodome==3.9.7
